refactor: dto file of post api
